新增: 物理表实体类及对应的mapper service

+++ fk-dataAccess/fk-dataAccess-pojo/src/main/java/com/fisk/dataaccess/dto/AppRegistrationDTO.java
@@ -29,7 +29,7 @@ public class AppRegistrationDTO extends BaseDTO {
 
     private String appDes;
 
-    private byte appType;
+    private int appType;
 
     private String appPrincipal;
